Move MapActivity to own package
